% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecks-package.R
\docType{package}
\name{tecks-package}
\alias{tecks-package}
\alias{tecks}
\title{Wrapper for Python's bibtexparser Library}
\description{
Wrapper for Python's bibtexparser Library
}
\author{
Scott Chamberlain \href{mailto:sckott@protonmail.com}{sckott@protonmail.com}
}
\keyword{package}
